Resolve generics before calling to_runtime_type (#51)
Summary: Since generic types are considered "known", and arguments to
generic types can be generic types themselves, it's difficult to
determine when it is actually safe to assume that a type is actually
fully known.  To enable this, this diff introduces a `resolve` function,
which returns a `CRef` that eventually resolves to the simplified type
with all generics fully elaborated.

---

<details open="true"><summary>Generated summary (powered by <a
href="https://app.graphite.dev">Graphite</a>)</summary>

>
> #### What changed
> - Added new methods to the Generic, GenericConstructor, SumGeneric,
CoerceGeneric, ExternalType and ConnectionType traits
> - Added new code to resolve_type, substitute and coerce_all functions
> - Added new code to resolve_generics and resolve_to_runtime_type
functions
>
> #### Impact
> This diff changes the way generics are resolved and how types are
casted to runtime types.
>
> #### Testing
> This diff could be tested by writing unit tests to ensure that the new
methods and functions behave as expected. Additionally, integration
tests could be written to ensure that the types are correctly resolved
and casted to runtime types.
</details>