Introduce the `for` operator (#78)
This change implements a new `foreach` operator that adds macro-like
capabilities to QueryScript, but in a type-safe and parser-friendly way.

Today, users reach for tools like DBT to produce SQL that operates over
an arbitrary set of fields. For example, imagine you want to write a
function that generates a SQL query that selects a few (inputted) fields
(`dimensions`). In Jinja, you might write something like:

```
SELECT
  {%- for dimension in dimensions -%}
    {{dimension}} / 10 {%- if not loop.last %} AS {{dimension}}, {% endif -%}
  {% endfor %}
FROM t
```

This is powerful because it allows you to analyze an arbitrary set of
inputs but cumbersome to write and brittle (notice the delicate
placement of the comma). Furthermore, unlike macro-based systems,
QueryScript has the additional responsibility of being type safe. That
means to support a pattern like this, we not only need to incorporate an
iterator into the syntax but also need to know exactly the type of its
output. We're not the only tool that supports patterns like this. In
fact,
[Clickhouse](https://clickhouse.com/docs/en/sql-reference/statements/select/#columns-expression)
and [DuckDB](https://duckdb.org/docs/sql/expressions/star#columns)
support a `COLUMNS(...)` operator which allows you to partially solve
for this pattern:

```
SELECT COLUMNS(`<regex_matching_dimensions>`) / 10 FROM t
```

Under the hood, `COLUMNS` explodes into `<field>/10` for each field
matching the provided regular expression. However, this syntax is not
fully general — for example you cannot straightforwardly specify each
field's alias. Furthermore, you can use `COLUMNS(...)` in places that
are syntactically valid but "not allowed" (try putting it in a `WHERE`
or `GROUP BY`).

This change introduces a new syntax that can be strategically placed in
supported spots, such as the project list, and is syntactically
unambiguous:

```
SELECT for dimension in dimensions {
  dimension / 10 AS dimension
}
FROM t
```

The loop is expanded at compile time, in terms of a constant array value
(`dimensions` in this example). Each instantiation of the loop body is
compiled and typechecked, as if you had handwritten it. And of course,
you do not need to do any fiddling with commas. The syntax also
naturally supports nested iteration:

As of this change, the `for` loop syntax is currently only available in
project lists, but it's trivial to extend support to other lists (`ORDER
BY`, `GROUP BY`). We can also reuse or tweak the syntax to support
joining together the body, for example `AND`-ing together each iteration
in the `WHERE` clause.