Isolate scope for projection terms (#85)
Before this change, we'd update the set of projection terms in the scope
as we compile the query. However, now that compilation is more complex,
we could end up in a state where we insert the projection term we're
currently compiling _before_ fully compiling that term, and therefore
deadlocking on `get_available_references()`.

This change reworks `compile_select_item()` to clone the scope for each
projection term, so that it has a frozen set of (pre-existing)
projection terms to compile against. It was a bit tricky to make this
work well for loops, which I did by adding a per-loop-body state object
that callers to `compile_foreach()` can create.

Although this change is quite difficult to verify, the deadlock
reproduced on an incoming change, which I've verified no longer
deadlocks with this change in place.