Support passing constants in functions to use in for loops (#97)
With this change, you can now pass in constant arrays to functions and
use them in for loops. This completes the major milestone of writing
functions that can take an arbitrary relation and list of fields as
input, and then operating over it.

We actually had most of the machinery in place to do this, but were
always passing in arguments as parameters (even for generic functions).
I changed `compile_fn_body` to accept an optional map of precompiled
arguments, which we substitute if present instead.

Furthermore, any function containing `for` expressions should go through
the deferred compilation codepath (where we stuff the expression body in
`Expr::Uncompiled`). For simplicity's sake, I changed _any_
expression-based function to now follow this codepath, whether or not it
has generics. Since this is mostly (or perhaps, purely) a performance
optimization, I think we can revisit it later to make sure that specific
cases (e.g. functions w/out for expressions) can still be compiled ahead
of time.