Use DuckDB to load files and support remote (e.g. HTTP) files (#100)
DuckDB has a lot of optimizations to process files efficiently,
including files that are compressed, remote, on S3, etc. This change
reworks how we load files to lean on this support instead of using Arrow
directly.

The crux of the change is renaming `eval` to `eval_lazy` and by default
returning a (potentially) lazily evaluated value. We take advantage of
this new mechanic to lazily load files and return a parameter with the
file name and format already parsed. In DuckDB, we intercept this
parameter and replace it with `read_csv_auto`, etc. In other engines,
these files will naturally just evaluate into relation parameters.