Add support for function parameters (#102)
While filling out support for the metrics playbook, I added support for
function parameters. This actually already worked, except for SQL
functions which we codegen by name. We were using the provided name of
the function (i.e. the parameter's name), not the SQL builtin's name, so
I simply updated `FnBody::SQLBuiltin` to keep the original name around
and use that.