Preliminary support for DBT (#105)
This change prototypes some building blocks (exposed through a python
library) that enable support for DBT. Specifically, we now support
compiling a schema into a set of queries and dependencies (with `ref()`
expressions) which can be consumed through a python library and invoked
through DBT.

While working through the DBT integration, I grappled with the fact that
`extern` is overloaded to mean "can import across modules" and "should
export when materializing/saving" (DBT or otherwise). Since this has
been a UX friction point for a while anyway, I just changed it to make
all decls public.